+++ probables/blooms/__init__.py
@@ -1,6 +1,6 @@
 ''' Bloom Filter based data structures '''
 from __future__ import (unicode_literals, absolute_import, print_function)
-from .bloom import (BloomFilter, BloomFilterOnDisk)
-from .countingbloom import (CountingBloomFilter)
+from . bloom import (BloomFilter, BloomFilterOnDisk)
+from . countingbloom import (CountingBloomFilter)
 
 __all__ = ['BloomFilter', 'BloomFilterOnDisk', 'CountingBloomFilter']