+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # PyProbables Changelog
 
+### Version 0.1.2:
+* Counting Cuckoo Filter
+    * Basic functionality: add, remove, check
+    * Expand
+    * Import / Export
+* Fix and tests for utility functions
+* Fix package build
+
 ### Version 0.1.1:
 * CuckooFilter
     * Import / Export functionality